# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v3/contents/generations/tasks": {
      "post": {
        "summary": "创建 ModelArk 视频任务",
        "deprecated": false,
        "description": "使用公开 ModelArk v3 合同创建视频任务。seedance-2.0-standard 与 seedance-2.0-fast 是独立的可变分辨率 SKU，两者仅支持普通 Link 资源，不支持真人素材；Fast 仅支持 480p/720p。飞彩 v2 固定分辨率 SKU 仍处于逐模型 size 与账单验证阶段，不在本 OpenAPI 中发布。",
        "operationId": "createModelArkVideoTask",
        "tags": [
          "视频生成/ModelArk格式"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelArkVideoCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "任务已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelArkVideoTask"
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "列出 ModelArk 视频任务",
        "deprecated": false,
        "operationId": "listModelArkVideoTasks",
        "tags": [
          "视频生成/ModelArk格式"
        ],
        "parameters": [
          {
            "name": "page_num",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任务列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelArkVideoTaskList"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "错误信息"
              },
              "type": {
                "type": "string",
                "description": "错误类型"
              },
              "param": {
                "type": "string",
                "description": "相关参数",
                "nullable": true
              },
              "code": {
                "type": "string",
                "description": "错误代码",
                "nullable": true
              }
            }
          }
        }
      },
...
      "ImageTaskError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ModelArkVideoCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model",
          "content"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "客户可见的 Seedance 模型 ID。一个已启用模型固定对应一个 Seedance Channel；实际可用性以 GET /v1/models 为准。"
          },
          "content": {
            "type": "array",
            "minItems": 1,
            "description": "统一 ModelArk V3 文本和媒体条目。媒体 URL 可使用 HTTP(S)、Data URL 或 asset://ast_*；具体模型能力由上线前技术审核和 Provider 错误确定。",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "type"
              ],
              "properties": {
                "type": {
                  "type": "string",
                  "enum": [
                    "text",
                    "image_url",
                    "video_url",
                    "audio_url"
                  ]
                },
                "text": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "first_frame",
                    "last_frame",
                    "reference_image",
                    "reference_video",
                    "reference_audio"
                  ]
                },
                "image_url": {
                  "$ref": "#/components/schemas/VideoMediaURL"
                },
                "video_url": {
                  "$ref": "#/components/schemas/VideoMediaURL"
                },
                "audio_url": {
                  "$ref": "#/components/schemas/VideoMediaURL"
                }
              }
            }
          },
          "duration": {
            "type": "integer",
            "minimum": -1,
            "maximum": 3600,
            "description": "视频时长；-1 表示智能时长，0 无效。模型自身范围可能更窄。"
          },
          "callback_url": {
            "type": "string",
            "format": "uri"
          },
          "resolution": {
            "type": "string",
            "description": "输出分辨率。可选值和必填性由当前已发布模型 capability 决定。"
          },
          "ratio": {
            "type": "string",
            "description": "画幅。可选值、默认值与 adaptive 支持由当前已发布模型 capability 决定。"
          },
          "service_tier": {
            "type": "string"
          },
          "generate_audio": {
            "type": "boolean",
            "description": "是否生成音频。是否允许显式 true/false 由当前已发布模型 capability 决定。"
          },
          "watermark": {
            "type": "boolean"
          },
          "return_last_frame": {
            "type": "boolean"
          },
          "execution_expires_after": {
            "type": "integer",
            "minimum": 3600,
            "maximum": 259200
          },
          "draft": {
            "type": "boolean"
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "type": {
                  "type": "string"
                }
              }
            }
          },
          "safety_identifier": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9
          },
          "frames": {
            "type": "integer",
            "minimum": 29,
            "maximum": 289,
            "description": "必须满足 25 + 4n；与 duration 同时存在时按官方语义优先。"
          },
          "seed": {
            "type": "integer",
            "minimum": -1,
            "maximum": 2147483647
          },
          "camera_fixed": {
            "type": "boolean"
          }
        }
      },
      "VideoMediaURL": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "ModelArkVideoTask": {
        "type": "object",
        "required": [
          "id",
          "model",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "content": {
            "type": "object",
            "properties": {
              "video_url": {
                "type": "string"
              },
              "last_frame_url": {
                "type": "string"
              }
            }
          },
          "seed": {
            "type": "integer",
            "format": "int64"
          },
          "resolution": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          },
          "frames": {
            "type": "integer"
          },
          "ratio": {
            "type": "string"
          },
          "generate_audio": {
            "type": "boolean"
          },
          "draft": {
            "type": "boolean"
          },
          "service_tier": {
            "type": "string"
          },
          "usage": {
            "type": "object",
            "properties": {
              "completion_tokens": {
                "type": "integer"
              },
              "total_tokens": {
                "type": "integer"
              }
            }
          },
          "error": {
            "$ref": "#/components/schemas/ImageTaskError"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ModelArkVideoTaskList": {
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelArkVideoTask"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}